+++ packages/web/src/main.tsx
@@ -31,8 +31,9 @@ import { Authenticator } from '@aws-amplify/ui-react';
 import UseCaseBuilderEditPage from './pages/useCaseBuilder/UseCaseBuilderEditPage.tsx';
 import App from './App.tsx';
 import UseCaseBuilderRoot from './UseCaseBuilderRoot.tsx';
-import UseCaseBuilderConsolePage from './pages/useCaseBuilder/UseCaseBuilderConsolePage.tsx';
 import UseCaseBuilderExecutePage from './pages/useCaseBuilder/UseCaseBuilderExecutePage.tsx';
+import UseCaseBuilderSamplesPage from './pages/useCaseBuilder/UseCaseBuilderSamplesPage.tsx';
+import UseCaseBuilderMyUseCasePage from './pages/useCaseBuilder/UseCaseBuilderMyUseCasePage.tsx';
 
 const ragEnabled: boolean = import.meta.env.VITE_APP_RAG_ENABLED === 'true';
 const ragKnowledgeBaseEnabled: boolean =
@@ -138,7 +139,11 @@ export const ROUTE_INDEX_USE_CASE_BUILDER = '/use-case-builder';
 const useCaseBuilderRoutes: RouteObject[] = [
   {
     path: ROUTE_INDEX_USE_CASE_BUILDER,
-    element: <UseCaseBuilderConsolePage />,
+    element: <UseCaseBuilderSamplesPage />,
+  },
+  {
+    path: `${ROUTE_INDEX_USE_CASE_BUILDER}/my-use-case`,
+    element: <UseCaseBuilderMyUseCasePage />,
   },
   {
     path: `${ROUTE_INDEX_USE_CASE_BUILDER}/new`,
@@ -147,10 +152,6 @@ const useCaseBuilderRoutes: RouteObject[] = [
   {
     path: `${ROUTE_INDEX_USE_CASE_BUILDER}/edit/:useCaseId`,
     element: <UseCaseBuilderEditPage />,
-  },
-  {
-    path: `${ROUTE_INDEX_USE_CASE_BUILDER}/chat/:chatId`,
-    element: <ChatPage />,
   },
   {
     path: `${ROUTE_INDEX_USE_CASE_BUILDER}/execute/:useCaseId`,
